Move method cleanData() to LoginUtils. Create cookiesFilePath variable.
