+++ src/test/java/utils/LoginUtils.java
@@ -3,7 +3,10 @@ package utils;
 import com.microsoft.playwright.Browser;
 import com.microsoft.playwright.BrowserContext;
 import com.microsoft.playwright.Page;
+import org.json.JSONObject;
 
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Paths;
 
 import static tests.BaseTest.waitForPageLoad;
@@ -12,6 +15,12 @@ import static utils.LoggerUtils.logError;
 
 public class LoginUtils {
     private static boolean tracingSave = false;
+    public static String cookiesFilePath = "src/test/resources/state.json";
+    private static String userToken;
+
+    public static String getUserToken() {
+        return userToken;
+    }
 
     public static boolean getTracingSave() {
         return tracingSave;
@@ -34,9 +43,9 @@ public class LoginUtils {
             login(page);
             if (page.url().equals(ProjectProperties.BASE_URL + TestData.HOME_END_POINT)) {
                 log("Login successful");
+                context.storageState(LoginUtils.storageStateOptions());
+                log("Cookies collected");
             }
-            context.storageState(LoginUtils.storageStateOptions());
-            log("Cookies collected");
         } catch (Exception e) {
             logError("ERROR: Unsuccessful login");
             tracingSave = true;
@@ -54,7 +63,7 @@ public class LoginUtils {
     public static BrowserContext.StorageStateOptions storageStateOptions() {
         return new BrowserContext
                 .StorageStateOptions()
-                .setPath(Paths.get("src/test/resources/state.json"));
+                .setPath(Paths.get(cookiesFilePath));
     }
 
     public static void login(Page page) {
@@ -69,5 +78,17 @@ public class LoginUtils {
         page.click("button[type='submit']");
 
         waitForPageLoad(page, TestData.HOME_END_POINT);
+    }
+
+    public static void parseUserToken() {
+        try {
+            String jsonString = new String(Files.readAllBytes(Paths.get(cookiesFilePath)));
+            JSONObject apiLoginResponseJSON = new JSONObject(jsonString);
+            String jsonValue = apiLoginResponseJSON.getJSONArray("origins").getJSONObject(0).getJSONArray("localStorage").getJSONObject(2).getString("value");
+            userToken = new JSONObject((new JSONObject(jsonValue)).getString("auth")).getString("accessToken");
+        }
+        catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 }
